+++ trixnity-serverserverapi/trixnity-serverserverapi-model/src/commonMain/kotlin/de/connect2x/trixnity/serverserverapi/model/SignedPersistentDataUnit.kt
@@ -1,9 +1,22 @@
 package de.connect2x.trixnity.serverserverapi.model
 
-import kotlinx.serialization.Contextual
 import de.connect2x.trixnity.core.model.events.PersistentDataUnit
 import de.connect2x.trixnity.core.model.keys.Signed
 
-typealias SignedPersistentDataUnit<T> = Signed<@Contextual PersistentDataUnit<T>, String>
-typealias SignedPersistentStateDataUnit<T> = Signed<@Contextual PersistentDataUnit.PersistentStateDataUnit<T>, String>
-typealias SignedPersistentMessageDataUnit<T> = Signed<@Contextual PersistentDataUnit.PersistentMessageDataUnit<T>, String>
+@Deprecated(
+    message = "Use Signed key type directly",
+    replaceWith = ReplaceWith("Signed<PersistentDataUnit<T>, String>")
+)
+typealias SignedPersistentDataUnit<T> = Signed<PersistentDataUnit<T>, String>
+
+@Deprecated(
+    message = "Use Signed key type directly",
+    replaceWith = ReplaceWith("Signed<PersistentDataUnit.PersistentStateDataUnit<T>, String>")
+)
+typealias SignedPersistentStateDataUnit<T> = Signed<PersistentDataUnit.PersistentStateDataUnit<T>, String>
+
+@Deprecated(
+    message = "Use Signed key type directly",
+    replaceWith = ReplaceWith("Signed<PersistentDataUnit.PersistentMessageDataUnit<T>, String>")
+)
+typealias SignedPersistentMessageDataUnit<T> = Signed<PersistentDataUnit.PersistentMessageDataUnit<T>, String>