+++ trixnity-serverserverapi/trixnity-serverserverapi-model/src/commonMain/kotlin/de/connect2x/trixnity/serverserverapi/model/federation/SendJoin.kt
@@ -1,20 +1,23 @@
 package de.connect2x.trixnity.serverserverapi.model.federation
 
-import io.ktor.resources.*
-import kotlinx.serialization.*
-import kotlinx.serialization.builtins.serializer
-import kotlinx.serialization.json.Json
 import de.connect2x.trixnity.core.HttpMethod
 import de.connect2x.trixnity.core.HttpMethodType.PUT
 import de.connect2x.trixnity.core.MatrixEndpoint
 import de.connect2x.trixnity.core.model.EventId
 import de.connect2x.trixnity.core.model.RoomId
+import de.connect2x.trixnity.core.model.events.PersistentDataUnit
 import de.connect2x.trixnity.core.model.events.PersistentDataUnit.PersistentStateDataUnit
 import de.connect2x.trixnity.core.model.events.m.room.MemberEventContent
 import de.connect2x.trixnity.core.model.keys.Signed
 import de.connect2x.trixnity.core.serialization.events.EventContentSerializerMappings
-import de.connect2x.trixnity.serverserverapi.model.SignedPersistentDataUnit
-import de.connect2x.trixnity.serverserverapi.model.SignedPersistentStateDataUnit
+import io.ktor.resources.*
+import kotlinx.serialization.Contextual
+import kotlinx.serialization.ExperimentalSerializationApi
+import kotlinx.serialization.KSerializer
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.builtins.serializer
+import kotlinx.serialization.json.Json
 
 /**
  * @see <a href="https://spec.matrix.org/v1.10/server-server-api/#put_matrixfederationv2send_joinroomideventid">matrix spec</a>
@@ -42,10 +45,10 @@ data class SendJoin(
     @Serializable
     data class Response(
         @SerialName("auth_chain")
-        val authChain: List<SignedPersistentDataUnit<*>>,
+        val authChain: List<Signed<@Contextual PersistentDataUnit<*>, String>>,
         @SerialName("event")
         val event: Signed<@Contextual PersistentStateDataUnit<MemberEventContent>, String>? = null,
         @SerialName("state")
-        val state: List<SignedPersistentStateDataUnit<*>>,
+        val state: List<Signed<@Contextual PersistentStateDataUnit<*>, String>>,
     )
 }